+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
  - Fix issues with SHT31 Smart Gadget and Ruuvitag Inputs ([#638](https://github.com/kizniche/mycodo/issues/638))
 
+### Features
+
+ - Increase Live page measurement query duration to fix the display of Input measurements
 
 ## 7.5.5 (2019-06-03)